reorganize src/components folder structure
