+++ client/src/components/index.js
@@ -1,3 +1,2 @@
-export { default as Button } from './Button';
-export { default as Tabs } from './Tabs';
-export { default as Tab } from './Tab';
+export { Button } from './Button';
+export { Tabs, Tab } from './Tabs';
